block

Train autoencoder and return blocks

block(namesA, namesB, trainNamesA, trainNamesB, model.path, dim.latent, dim.encode, dim.decode,
max.length, num.encode.layers, num.decode.layers, batch.size, epochs, lr, validation.split, save.dir, earlystop, earlystop.patience,
tensorboard, tensorboard.runid)

Arguments

dim.latent
Number of dimensions to use in the latent vector

dim.encode
Number of hidden units to use in each of the encoding layers

dim.decode
Number of hidden units to use in each of the decoding layers

max.length
Maximum length of each name

num.encode.layers
Number of encoding layers to use

num.decode.layers
Number of decoding layers to use

batch.size
Batch size of model training

epochs
Number of epochs to train for

lr
Learning rate of model. This model is configured to use to the Adam optimizer.

validation.split
Percentage of data to save for validation

save.dir
Directory to save the trained encoder

earlystop
TRUE if stopping early when validation loss is no longer decreasing, if FALSE then train for all epochs

earlystop.patience
Number of epochs to wait while validation loss does not decrease before stopping training early

tensorboard
TRUE if tensorboard metrics are to be recorded. Logs are recorded in the /tmp/ directory

tensorboard.runid
Unique identifier for the run to separate tensorboard logs

namesA
vector of names from dataset A

namesB
vector of names from dataset B

trainNamesA
vector of names from dataset A that are known matches

trainNamesB
vector of names from dataset B that are known matches in the same order as trainNamesA

Value

block returns a nested list with the following items

Nested list of blocks where each block is one encoding. Sub-lists are of indices of namesA and namesB belonging to each block

Examples

# NOT RUN {
# Sample names 
namesA <- c('amy', 'john', 'cathy', 'jim')
namesB <- c('amy', 'cathy', 'jim', 'clara', 'jon', 'dave')

trainNamesA <- c('amy', 'john')
trainNamesB <- c('amy', 'jon')

blocks <- block(namesA, namesB, trainNamesA, trainNamesB)
# }